Start tracking treeshakability
This diff adds size snapshot entry with esm bundle which allows to see
that some part of the code is not treeshakable.

This project is used by material-ui which provides only path imports
which looks very scary (a lot of import statement). So this project
should be treeshakable to let its users safely consumer from single
entry point and do not care about big bundle size.