fix: referenceNode memory leak
Instead of storing referenceNode in state, store it in a class property. This is done to enable the component to cleanup referenceNode’s value during unmounting.

I also added unit tests to account for this change, but that resulted in me having to update some dependencies and snapshots to get unit tests to pass.